MARK: This is some serious bug here so I guess I'll document it here Whether using self in animation is still to be discussed, should check later
POSSIBLE MEMORY LEAK, DUE TO AUTOMATIC REFERENCE COUNT
LEADING TO THE CUSTOMARVIEW NOT BEING DISMANTLED PROPERLY
HOLDING THE TEXTURE/MATERIAL
AND THE NEXT ARVIEW WILL NOT BE ALBE TO USE THOSE PROPERLY

SOLUTION: USE WEAK SELF

It's worth noting that, ARC will lead to possible memory reference loop
Which is quite easy to happen in closure with reference to self if you're not careful
Problem is, memory leak isn't just memory leak. Metal seems to handle textures in a way that
Only one ARSession/ARView can reference it at the same time
So if your old ARView is not destroyed properly, the next one will suffer!
Leading to some extremely unsatisfying and "undebugable" senarios
In my case, the APP crashes immediately in debug mode, and displays inproperly/
refuses to handle gestures, without debugging.